Extract /zotero/* routes into prisma/server/zotero_routes.py
Third of app.py's ~14 route domains to split out (F4). Same
build_zotero_router(deps) -> APIRouter factory pattern as
sync_routes.py/notes_routes.py/streams_routes.py: getter callables for
_vault/_zotero/_indexer (all rebound at runtime by /reload/* endpoints).

Moves all 6 /zotero routes (status, collections, stats, sync-pending,
items, import/{key}) plus their request/response models
(ZoteroStatsResponse, SyncPendingResponse) and the two PDF-fetch helpers
(_fetch_pdf_from_url, _pdf_bytes_to_md) that only zotero_import used.

The existing tests/unit/server/test_zotero_routes.py (full-app-singleton
style, monkeypatching app_module._zotero) needed no changes -- the closure
getters mean app_module._zotero is still the thing that actually answers
requests, just reached one level indirected through the router now.
Extended it with 6 new tests for /zotero/status, /zotero/collections, and
/zotero/import/{key} -- the latter had zero coverage anywhere despite real
branching logic (existing-import dedup, PDF-vs-abstract-fallback body,
citekey creation) -- built via build_zotero_router in isolation (bare
FastAPI app + tmp_path VaultService), matching test_notes_routes.py's
approach, since import writes real files and the module-level `client`
fixture talks to the full app's real (non-tmp) _vault.

app.py: 1894 -> 1700 lines.

Found during a modularization re-audit against
software-engineering-quality-aspects.md (F4, third extraction --
admin_routes.py/search_routes.py still to come).
